Fixes for building on MinGW GCC on Windows 7.

+++ libcola/colafd.cpp
@@ -21,6 +21,7 @@
  *
 */
 
+#include <cstring>
 #include <vector>
 #include <cmath>
 #include <limits>
@@ -1221,7 +1222,8 @@ void ConstrainedFDLayout::outputInstanceToSVG(std::string instanceName)
         {
             if (G[i][j] == 1)
             {
-                fprintf(fp, "    es.push_back(std::make_pair(%lu, %lu));\n", i, j);
+                fprintf(fp, "    es.push_back(std::make_pair(%lu, %lu));\n",
+                        (unsigned long) i, (unsigned long) j);
             }
         }
     }
